requirements.txt
boto3>=1.26.0
azure-storage-blob>=12.14.0
